Split `DeliveryArea` tests into per-type files
`src/.../grid/_delivery_area.py` exposes five public symbols
(`EnergyMarketCodeType`, `BaseDeliveryArea`, `DeliveryArea`,
`InvalidDeliveryArea`, `InvalidDeliveryAreaError`), so the test file
grew quite big.

This commit splits the test file following the pattern:

    tests/grid/test_delivery_area.py
      ->  tests/grid/_delivery_area/test_<thing>.py

Test bodies and assertions are unchanged; only their location and their
names change. Test function names are stripped of the redundant
module-name prefix per the same guideline (e.g.
`test_invalid_delivery_area_error_default_message` -> `test_default_message`
inside `test_invalid_delivery_area_error.py`).
